began moved causal stuff from model to inference
